refactor: component audit — adopt admin_page wrapper, split oversized components
- Replace duplicated header+tab_nav block in all 8 admin templates with
  the existing admin_page/1 component wrapper
- Fix stale FlareSolverr "Configure" link (/admin/config?tab=general →
  /admin/config/settings)
- Split admin_indexers_live/components.ex (1,252→398 LOC) by extracting
  modals to modals.ex (876 LOC)
- Split admin_quality_profiles_live/components.ex (1,063→187 LOC) by
  extracting modals to modals.ex (884 LOC)
