+++ lib/mydia_web/live/admin_download_clients_live/index.html.heex
@@ -1,22 +1,9 @@
 <Layouts.app {assigns}>
-  <div>
-    <div class="flex flex-col md:flex-row md:items-center md:justify-between gap-4 mb-6">
-      <div>
-        <h1 class="text-3xl font-bold">Configuration</h1>
-        <p class="text-base-content/70 mt-1">
-          System status, application settings, and configuration management
-        </p>
-      </div>
-    </div>
-
-    <MydiaWeb.AdminComponents.tab_nav active_tab={@active_tab} />
-
-    <div class="bg-base-100">
-      <MydiaWeb.AdminDownloadClientsLive.Components.download_clients_tab
-        download_clients={@download_clients}
-        client_health={@client_health}
-      />
-    </div>
+  <.admin_page active_tab={@active_tab}>
+    <MydiaWeb.AdminDownloadClientsLive.Components.download_clients_tab
+      download_clients={@download_clients}
+      client_health={@client_health}
+    />
 
     <%!-- Download Client Modal --%>
     <%= if assigns[:show_download_client_modal] do %>
@@ -26,5 +13,5 @@
         testing_download_client_connection={assigns[:testing_download_client_connection] || false}
       />
     <% end %>
-  </div>
+  </.admin_page>
 </Layouts.app>
