+++ lib/mydia_web/live/admin_settings_live/index.html.heex
@@ -1,21 +1,8 @@
 <Layouts.app {assigns}>
-  <div>
-    <div class="flex flex-col md:flex-row md:items-center md:justify-between gap-4 mb-6">
-      <div>
-        <h1 class="text-3xl font-bold">Configuration</h1>
-        <p class="text-base-content/70 mt-1">
-          System status, application settings, and configuration management
-        </p>
-      </div>
-    </div>
-
-    <MydiaWeb.AdminComponents.tab_nav active_tab={@active_tab} />
-
-    <div class="bg-base-100">
-      <Components.general_settings_tab
-        config_settings_with_sources={@config_settings_with_sources}
-        crash_report_stats={@crash_report_stats}
-      />
-    </div>
-  </div>
+  <.admin_page active_tab={@active_tab}>
+    <Components.general_settings_tab
+      config_settings_with_sources={@config_settings_with_sources}
+      crash_report_stats={@crash_report_stats}
+    />
+  </.admin_page>
 </Layouts.app>
