+++ lib/mydia_web/live/admin_system_live/index.html.heex
@@ -1,36 +1,23 @@
 <Layouts.app {assigns}>
-  <div>
-    <div class="flex flex-col md:flex-row md:items-center md:justify-between gap-4 mb-6">
-      <div>
-        <h1 class="text-3xl font-bold">Configuration</h1>
-        <p class="text-base-content/70 mt-1">
-          System status, application settings, and configuration management
-        </p>
+  <.admin_page active_tab={@active_tab}>
+    <MydiaWeb.AdminSystemLive.Components.status_tab
+      system_info={@system_info}
+      database_info={@database_info}
+      library_paths_count={@library_paths_count}
+      download_clients_count={@download_clients_count}
+      indexers_count={@indexers_count}
+      active_sessions={@active_sessions}
+      active_jobs={@active_jobs}
+      recent_activity={@recent_activity}
+    />
+    <%!-- FlareSolverr Status (External Services) --%>
+    <div class="px-4 sm:px-6 pb-6">
+      <div class="divider">External Services</div>
+      <div class="grid grid-cols-1 lg:grid-cols-2 gap-6 sm:gap-8">
+        <MydiaWeb.FlareSolverrStatusComponent.flaresolverr_status_card flaresolverr_status={
+          @flaresolverr_status
+        } />
       </div>
     </div>
-
-    <MydiaWeb.AdminComponents.tab_nav active_tab={@active_tab} />
-
-    <div class="bg-base-100">
-      <MydiaWeb.AdminSystemLive.Components.status_tab
-        system_info={@system_info}
-        database_info={@database_info}
-        library_paths_count={@library_paths_count}
-        download_clients_count={@download_clients_count}
-        indexers_count={@indexers_count}
-        active_sessions={@active_sessions}
-        active_jobs={@active_jobs}
-        recent_activity={@recent_activity}
-      />
-      <%!-- FlareSolverr Status (External Services) --%>
-      <div class="px-4 sm:px-6 pb-6">
-        <div class="divider">External Services</div>
-        <div class="grid grid-cols-1 lg:grid-cols-2 gap-6 sm:gap-8">
-          <MydiaWeb.FlareSolverrStatusComponent.flaresolverr_status_card flaresolverr_status={
-            @flaresolverr_status
-          } />
-        </div>
-      </div>
-    </div>
-  </div>
+  </.admin_page>
 </Layouts.app>
